+++ src/Rector/ORM/ListFilterToArrayRector.php
@@ -39,7 +39,7 @@ class ListFilterToArrayRector extends AbstractRector implements DocumentedRuleIn
             return null;
         }
 
-        if (!$this->isObjectType($node->var, new \PHPStan\Type\ObjectType('SilverStripe\ORM\DataList')) &&
+        if (!$this->isObjectType($node->var, new \PHPStan\Type\ObjectType(\SilverStripe\ORM\DataList::class)) &&
             !$this->isObjectType($node->var, new \PHPStan\Type\ObjectType('SilverStripe\ORM\ArrayList'))
         ) {
             return null;