+++ stubs/SilverStripe/Forms/FieldList.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SilverStripe\Forms;
+
+class FieldList
+{
+    public function dataFields(): array
+    {
+        return [];
+    }
+
+    public function getDataFields(): array
+    {
+        return [];
+    }
+}