+++ stubs/SilverStripe/ORM/EagerLoadedList.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SilverStripe\ORM;
+
+class EagerLoadedList
+{
+    public function getIDList(): array
+    {
+        return [];
+    }
+
+    public function column($column = 'ID'): array
+    {
+        return [];
+    }
+}